python/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==1.10.7
requests==2.28.2
pymongo==4.3.3